elk_flask/elk/app/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
python-json-logger==2.0.4
python-logstash==0.4.8
